+++ src/Resolvers/LocaleResolver.php
@@ -4,7 +4,7 @@ namespace Goodcat\L10n\Resolvers;
 
 use Illuminate\Http\Request;
 
-interface PreferredLocaleResolver
+interface LocaleResolver
 {
     public function resolve(Request $request): ?string;
 }