+++ src/Resolvers/UserLocale.php
@@ -5,7 +5,7 @@ namespace Goodcat\L10n\Resolvers;
 use Illuminate\Contracts\Translation\HasLocalePreference;
 use Illuminate\Http\Request;
 
-class UserPreferredLocale implements PreferredLocaleResolver
+class UserLocale implements LocaleResolver
 {
     public function resolve(Request $request): ?string
     {